+++ src/index.ts
@@ -13,7 +13,7 @@ export type RuleFunction = 'no-function' | 'no-sub' | 'auto' | 'off';
 export type RuleAAComma = 'always' | 'no-dangling' | 'never' | 'off';
 export type RuleTypeAnnotations = 'all' | 'return' | 'args' | 'off';
 export type RuleEolLast = 'always' | 'never' | 'off';
-export type RuleColorFormat = 'hash' | 'zero-x' | 'never' | 'off';
+export type RuleColorFormat = 'hashHex' | 'quotedNumericHex' | 'never' | 'off';
 export type RuleColorCase = 'upper' | 'lower' | 'off';
 export type RuleColorAlpha = 'always' | 'allowed' | 'never' | 'off';
 export type RuleColorAlphaDefaults = 'allowed' | 'only-hidden' | 'never' | 'off';
