+++ src/util.ts
@@ -1,12 +1,11 @@
 import { parse } from 'jsonc-parser';
 import * as minimatch from 'minimatch';
-import { BsLintConfig, BsLintRules, RuleSeverity, BsLintSeverity, RuleColorFormat, RuleColorCase, RuleColorAlpha, RuleColorAlphaDefaults, RuleColorCertCompliant } from './index';
+import { BsLintConfig, BsLintRules, RuleSeverity, BsLintSeverity } from './index';
 import { readFileSync, existsSync } from 'fs';
 import * as path from 'path';
-import { Program, BscFile, DiagnosticSeverity, BsDiagnostic, Range } from 'brighterscript';
+import { Program, BscFile, DiagnosticSeverity } from 'brighterscript';
 import { applyFixes, ChangeEntry, TextEdit } from './textEdit';
 import { addJob } from './Linter';
-import { messages } from './plugins/codeStyle/diagnosticMessages';
 
 export function getDefaultRules(): BsLintConfig['rules'] {
     return {
@@ -41,105 +40,6 @@ export function normalizeConfig(options: BsLintConfig) {
     };
     const projectConfig = mergeConfigs(loadConfig(options), { rules: options.rules });
     return mergeConfigs(baseConfig, projectConfig);
-}
-
-export function validateColorStyle(text: string, range: Range, diagnostics: (Omit<BsDiagnostic, 'file'>)[]) {
-    const colorHashRegex = /#[0-9A-Fa-f]{6}/g;
-    const colorHashAlphaRegex = /#[0-9A-Fa-f]{8}/g;
-    const colorZeroXRegex = /0x[0-9A-Fa-f]{6}/g;
-    const colorZeroXAlphaRegex = /0x[0-9A-Fa-f]{8}/g;
-    const colorHashMatches = text.match(colorHashRegex);
-    const colorHashAlphaMatches = text.match(colorHashAlphaRegex);
-    const colorZeroXMatches = text.match(colorZeroXRegex);
-    const colorZeroXAlphaMatches = text.match(colorZeroXAlphaRegex);
-
-    if (colorFormat === 'hash') {
-        if (colorZeroXMatches !== null) {
-            diagnostics.push(messages.expectedColorFormat(range));
-        }
-        validateColorCase(colorHashMatches, range, diagnostics, colorCase, colorFormat);
-        validateColorAlpha(colorHashAlphaMatches, colorHashMatches, colorZeroXMatches, range, diagnostics, alpha, alphaDefaults);
-        validateColorCertCompliance(colorHashMatches, range, diagnostics, colorFormat, certCompliant);
-
-    } else if (colorFormat === 'zero-x') {
-        if (colorHashMatches !== null) {
-            diagnostics.push(messages.expectedColorFormat(range));
-        }
-        validateColorCase(colorZeroXMatches, range, diagnostics, colorCase, colorFormat);
-        validateColorAlpha(colorZeroXAlphaMatches, colorHashMatches, colorZeroXMatches, range, diagnostics, alpha, alphaDefaults);
-        validateColorCertCompliance(colorZeroXMatches, range, diagnostics, colorFormat, certCompliant);
-
-    } else if (colorFormat === 'never') {
-        if (colorZeroXMatches !== null || colorHashMatches !== null) {
-            diagnostics.push(messages.expectedColorFormat(range));
-        }
-    }
-}
-
-function validateColorAlpha(alphaMatches: RegExpMatchArray, hashMatches: RegExpMatchArray, zeroXMatches: RegExpMatchArray, range: Range, diagnostics: (Omit<BsDiagnostic, 'file'>)[], alpha: RuleColorAlpha, alphaDefaults: RuleColorAlphaDefaults) {
-    const validateColorAlpha = (alpha === 'never' || alpha === 'always' || alpha === 'allowed');
-    if (validateColorAlpha) {
-        if (alpha === 'never' && alphaMatches !== null) {
-            diagnostics.push(messages.expectedColorAlpha(range));
-        }
-        if ((alpha === 'always' && alphaMatches === null) && (hashMatches !== null || zeroXMatches !== null)) {
-            diagnostics.push(messages.expectedColorAlpha(range));
-        }
-        if ((alphaDefaults === 'never' || alphaDefaults === 'only-hidden') && alphaMatches !== null) {
-            for (let i = 0; i < alphaMatches.length; i++) {
-                const colorHashAlpha = alphaMatches[i];
-                const alphaValue = colorHashAlpha.slice(-2).toLowerCase();
-                if (alphaValue === 'ff' || (alphaDefaults === 'never' && alphaValue === '00')) {
-                    diagnostics.push(messages.expectedColorAlphaDefaults(range));
-                }
-            }
-        }
-    }
-}
-
-function validateColorCase(matches: RegExpMatchArray, range: Range, diagnostics: (Omit<BsDiagnostic, 'file'>)[], colorCase: RuleColorCase, colorFormat: RuleColorFormat) {
-    const validateColorCase = colorCase === 'upper' || colorCase === 'lower';
-    if (validateColorCase && matches !== null) {
-        let colorValue = matches[0];
-        const charsToStrip = (colorFormat === 'hash') ? 1 : 2;
-        colorValue = colorValue.substring(charsToStrip);
-        for (let i = 0; i < colorValue.length; i++) {
-            const char = colorValue.charAt(i);
-            if (colorCase === 'lower' && char === char.toUpperCase() && char !== char.toLowerCase()) {
-                diagnostics.push(messages.expectedColorCase(range));
-                break;
-            }
-            if (colorCase === 'upper' && char === char.toLowerCase() && char !== char.toUpperCase()) {
-                diagnostics.push(messages.expectedColorCase(range));
-                break;
-            }
-        }
-    }
-}
-
-function validateColorCertCompliance(matches: RegExpMatchArray, range: Range, diagnostics: (Omit<BsDiagnostic, 'file'>)[], colorFormat: RuleColorFormat, certCompliant: RuleColorCertCompliant) {
-    const validateCertCompliant = certCompliant === 'always';
-    if (validateCertCompliant && matches !== null) {
-        const BROADCAST_SAFE_BLACK = '161616';
-        const BROADCAST_SAFE_WHITE = 'DBDBDB';
-        const MAX_BLACK_LUMA = getColorLuma(BROADCAST_SAFE_BLACK);
-        const MAX_WHITE_LUMA = getColorLuma(BROADCAST_SAFE_WHITE);
-        let colorValue = matches[0];
-        const charsToStrip = (colorFormat === 'hash') ? 1 : 2;
-        colorValue = colorValue.substring(charsToStrip);
-        const colorLuma = getColorLuma(colorValue);
-        if (colorLuma > MAX_WHITE_LUMA || colorLuma < MAX_BLACK_LUMA) {
-            diagnostics.push(messages.colorCertCompliance(range));
-        }
-    }
-}
-
-function getColorLuma(value: string) {
-    const rgb = parseInt(value, 16); // Convert rrggbb to decimal
-    const red = (rgb >> 16) & 0xff;
-    const green = (rgb >> 8) & 0xff;
-    const blue = (rgb >> 0) & 0xff;
-    return 0.2126 * red + 0.7152 * green + 0.0722 * blue; // Per ITU-R BT.709
 }
 
 export function mergeConfigs(a: BsLintConfig, b: BsLintConfig): BsLintConfig {
